Move ValidationErrorFactory to domain
