+++ src/datasources/config-api/config-api.module.ts
@@ -2,7 +2,7 @@ import { Global, Module } from '@nestjs/common';
 import { ConfigApi } from './config-api.service';
 import { CacheFirstDataSourceModule } from '../cache/cache.first.data.source.module';
 import { IConfigApi } from '../../domain/interfaces/config-api.interface';
-import { ValidationErrorFactory } from '../errors/validation-error-factory';
+import { ValidationErrorFactory } from '../../domain/errors/validation-error-factory';
 import { JsonSchemaService } from '../../common/schemas/json-schema.service';
 
 @Global()
