+++ src/datasources/exchange-api/exchange-api.module.ts
@@ -2,7 +2,7 @@ import { Global, Module } from '@nestjs/common';
 import { HttpErrorFactory } from '../errors/http-error-factory';
 import { ExchangeApi } from './exchange-api.service';
 import { IExchangeApi } from '../../domain/interfaces/exchange-api.interface';
-import { ValidationErrorFactory } from '../errors/validation-error-factory';
+import { ValidationErrorFactory } from '../../domain/errors/validation-error-factory';
 import { JsonSchemaService } from '../../common/schemas/json-schema.service';
 
 @Global()
