+++ src/datasources/transaction-api/transaction-api.module.ts
@@ -2,7 +2,7 @@ import { Global, Module } from '@nestjs/common';
 import { TransactionApiManager } from './transaction-api.manager';
 import { CacheFirstDataSourceModule } from '../cache/cache.first.data.source.module';
 import { ITransactionApiManager } from '../../domain/interfaces/transaction-api.manager.interface';
-import { ValidationErrorFactory } from '../errors/validation-error-factory';
+import { ValidationErrorFactory } from '../../domain/errors/validation-error-factory';
 import { JsonSchemaService } from '../../common/schemas/json-schema.service';
 
 @Global()
